requirements.txt
django==4.2.11
djangorestframework==3.14.0
celery==5.3.6
django-celery-results==2.5.1
redis==5.0.3
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
groq>=0.9.0
requests==2.31.0
beautifulsoup4==4.12.3
duckduckgo-search==5.3.1
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
